// src/schemaValidations/role.schema.ts
import { z } from 'zod'
import { PermissionSchema } from './permission.schema'

export const RoleSchema = z.object({
    id: z.number(),
    name: z.string().max(500),
    description: z.string(),
    isActive: z.boolean().default(true),
    createdById: z.number().nullable(),
    updatedById: z.number().nullable(),
    deletedById: z.number().nullable(),
    deletedAt: z.iso.datetime().nullable(),
    createdAt: z.iso.datetime(),
    updatedAt: z.iso.datetime(),
})

export const RolePermissionsSchema = RoleSchema.extend({
    permissions: z.array(PermissionSchema),
})

export type RoleType = z.infer<typeof RoleSchema>
export type RolePermissionsType = z.infer<typeof RolePermissionsSchema>